run.py
#!/usr/bin/env python3
"""
Audio Enhancement Application Runner
"""
import uvicorn
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

if __name__ == "__main__":
    # Configuration
    host = os.getenv("HOST", "0.0.0.0")
    port = int(os.getenv("PORT", 8000))
    reload = os.getenv("RELOAD", "True").lower() == "true"
    
    print("🎵 Starting Audio Enhancement Server...")
    print(f"📍 Server: http://{host}:{port}")
    print(f"🔄 Auto-reload: {reload}")
    
    # Run the application
    uvicorn.run(
        "app.main:app",
        host=host,
        port=port,
        reload=reload,
        log_level="info"
    )