requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Audio processing 
torch
torchaudio
speechbrain
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Additional dependencies for HF Spaces
requests==2.31.0
numpy==1.24.3